app/requirements.txt
fastapi[standard]==0.115.12
opencv-python==4.11.0.86
pydantic==2.11.7
starlette==0.46.2
uvicorn==0.34
sqlalchemy==2.0.41
pymysql==1.1.1
cryptography==45.0.4
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
python-jose[cryptography]==3.5.0